requirements/common.txt
Django==1.5.5
Pillow==2.4.0
South==0.8.4
argparse==1.2.1
django-classy-tags==0.5
django-configurations==0.8
django-extensions==1.3.3
django-filer==0.9.5
django-ignoretests==0.3.2
django-mptt==0.5.5
django-polymorphic==0.5.3
django-sekizai==0.7
django-suit==0.2.7
easy-thumbnails==1.5
psycopg2==2.5.2
six==1.6.1
wsgiref==0.1.2
